Cast RAND_MAX to float to avoid Clang warning.
implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648

+++ base/utils.c
@@ -3294,7 +3294,7 @@ int check_for_nagios_updates(int force, int reschedule) {
 		/* we didn't do an update, so calculate next possible update time */
 		if(do_check == FALSE) {
 			next_check = last_update_check + BASE_UPDATE_CHECK_INTERVAL;
-			next_check = next_check + (unsigned long)(((float)randnum / RAND_MAX) * UPDATE_CHECK_INTERVAL_WOBBLE);
+			next_check = next_check + (unsigned long)(((float)randnum / (float)RAND_MAX) * UPDATE_CHECK_INTERVAL_WOBBLE);
 			}
 
 		/* we tried to check for an update */
@@ -3303,13 +3303,13 @@ int check_for_nagios_updates(int force, int reschedule) {
 			/* api query was okay */
 			if(api_result == OK) {
 				next_check = current_time + BASE_UPDATE_CHECK_INTERVAL;
-				next_check += (unsigned long)(((float)randnum / RAND_MAX) * UPDATE_CHECK_INTERVAL_WOBBLE);
+				next_check += (unsigned long)(((float)randnum / (float)RAND_MAX) * UPDATE_CHECK_INTERVAL_WOBBLE);
 				}
 
 			/* query resulted in an error - retry at a shorter interval */
 			else {
 				next_check = current_time + BASE_UPDATE_CHECK_RETRY_INTERVAL;
-				next_check += (unsigned long)(((float)randnum / RAND_MAX) * UPDATE_CHECK_RETRY_INTERVAL_WOBBLE);
+				next_check += (unsigned long)(((float)randnum / (float)RAND_MAX) * UPDATE_CHECK_RETRY_INTERVAL_WOBBLE);
 				}
 			}
